requirements.txt
python-telegram-bot==20.8
pandas
openpyxl
requests==2.32.4